A little late night fun.
Did some cleaning of onCreate, implemented the getNumConnections function (and it builds and runs without
crashing, so I'll call that a success for now), and am still learning.

Next step is to implement the ability to write to my file (which I will be able to use for testing the
changes that I made tonight) and get ready to do some BlueTooth implementation that would make Harold
proud.