Tested BT and made some changes.
For some reason I was attempting to track whether we were using BT by a global
variable, which caused a bug when switching BT on and off while app was running.

Fixed this bug and added some clarity for what else I need to do to get this
working in a few weeks.